src/requirements.txt
Django==3.0.8
djangorestframework==3.11.0
django-environ==0.4.5
pandas==1.1.1
psycopg2-binary==2.8.5
